agent-runtime/outputs/backend/requirements.txt
httpx==0.27.0
gspread==6.1.2
google-auth==2.29.0
google-api-python-client==2.128.0
pydantic==2.7.1
pytz==2024.1
